Check receptor signaling downstream

Check receptor signaling downstream.

checkReceptorSignaling(
  ds,
  lr,
  method = c("reactome-GOBP", "reactome", "GOBP"),
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  restrict.pw = NULL,
  with.complex = TRUE
)

Arguments

ds
A BulkSignalR data set.

lr
A ligand-receptor analysis such as returned by getCorrelatedLR().

method
Which pathway reference should be used ('reactome' for Reactome,'GOBP' for GO Biological Process, or 'Reactome-GOBP' for both).

max.pw.size
Maximum pathway size to consider from the pathway reference.

min.pw.size
Minimum pathway size to consider from the pathway reference.

min.positive
Minimum number of target genes to be found in a given pathway.

restrict.pw
A list of pathway IDs to restrict the application of the function.

with.complex
A logical indicating whether receptor co-complex members should be included in the target genes.

Value

A ligand-receptor analysis extending lr content with a table reporting all the ligand-receptor pairs provided in lr along with the pathways found and data about target gene correlations with the receptor. Those data are the lists (in semi-colon-separated format) of target genes and their Spearman correlations with the receptor. The lists are sorted according to the correlation coefficient.

@details In a pathway of the reference, i.e., a Reactome pathway or the genes of a GOBP term, the target genes are the genes coding for proteins forming a complex with the receptor and the genes in the pathway downstream the receptor, which are given as regulated by the pathway. If with.complex is set to FALSE, then only the regulated genes are considered. Participation to a complex and being regulated as well as the pathway directed topologies are defined by Reactome and KEGG pathways as provided by PathwayCommons.

The maximum pathway size is used to limit the redundancy inherent to GOBP and Reactome. The minimum pathway size is used to avoid overspecific, noninformative results.

Details

Assess the existence of correlations between a receptor, part of a ligand-receptor pair, and genes coding for proteins forming a complex with the receptor or genes regulated by the receptor downstream signaling.

Examples

print('checkReceptorSignaling')
#> [1] "checkReceptorSignaling"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
#> [1] "REACTOME-GOBP"