Internal function to check receptor signaling downstream

Assess the existence of correlations between a receptor, part of a ligand-receptor pair, and genes coding for proteins forming a complex with the receptor or genes regulated by the receptor downstream signaling.

.checkReceptorSignaling(
  ds,
  lr,
  reference = c("REACTOME-GOBP", "REACTOME", "GOBP"),
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  restrict.pw = NULL,
  with.complex = TRUE
)

Arguments

ds

A BSRDataModel object.

lr

A table as returned by .getCorrelatedLR().

reference

Which pathway reference should be used ("REACTOME" for Reactome, "GOBP" for GO Biological Process, or "REACTOME-GOBP" for both).

max.pw.size

Maximum pathway size to consider from the pathway reference.

min.pw.size

Minimum pathway size to consider from the pathway reference.

min.positive

Minimum number of target genes to be found in a given pathway.

restrict.pw

A list of pathway IDs to restrict the application of the function.

with.complex

A logical indicating whether receptor co-complex members should be included in the target genes.

Value

A data frame extending lr content with the pathways found to contain the receptors and data about target gene correlations with those receptors. Strings in semi-colon-separated format are used to report target genes and their Spearman correlations with the receptor in the data frame. The target genes are sorted according to the correlation coefficient. In a pathway of the reference, i.e., a Reactome pathway or the genes of a GOBP term, the target genes are the genes coding for proteins forming a complex with the receptor and the genes in the pathway downstream the receptor, which are given as regulated by the pathway. If with.complex is set to FALSE, then only the regulated genes are considered. Participation to a complex and being regulated as well as the pathway directed topologies are defined by Reactome and KEGG pathways as provided by PathwayCommons. The maximum pathway size is used to limit the redundancy inherent to GOBP and Reactome. The minimum pathway size is used to avoid overspecific, noninformative results.