Heatmap function for LR scores and additional data

Generate a stack of heatmaps. The top heatmap represents ligand-receptor gene signature scores, while the bottom heatmap represents the second, user-chosen data set to feature along with gene signatures.

dualHeatmap(
  mat.c,
  mat.e,
  file.name = NULL,
  dend.row = NULL,
  dend.spl = NULL,
  dend.e = NULL,
  cols = NULL,
  cols.e = NULL,
  width,
  height = 6,
  pointsize = 4,
  cut.p = 0.01,
  vert.p = 0.9,
  row.names.1 = TRUE,
  row.names.2 = TRUE
)

Arguments

mat.c
A matrix with the signature scores such as output by scoreLRGeneSignatures().

mat.e
A second matrix to feature below mat.c.

file.name
A PDF file name.

dend.row
A precomputed row dendrogram for mat.c.

dend.spl
A precompute sample (column) dendrogram.

dend.e
A precomputed row dendrogram for mat.e.

cols
A vector of colors to use for the heatmap of mat.c.

cols.e
A vector of colors to use for the heatmap of mat.e.

width
PDF width.

height
PDF height.

pointsize
PDF pointsize.

cut.p
Proportion of top and bottom values for thresholding.

vert.p
Proportion of total height dedicated to the top heatmap.

row.names.1
A logical to turn on/off the display of row (gene) names in the top heatmap.

row.names.2
A logical to turn on/off the display of row (gene) names in the bottom heatmap.

Value

A heatmap. Since heatmap plotting tend to be slow on the screen, it is advisable to provide a PDF file name and plot in a file (much faster).

Details

Extreme values (top and bottom) can be replaced by global quantiles at cut.p and 1-cut.p to avoid color scales shrunk by a few outliers.

This is a convenience function that relies on the ComplexHeatmap package to propose a simple way of representing signature scores. If more advanced features are needed or more graphic parameters should be controled, users should implement their own function, e.g., strarting from the code of dualHeatmap().

Examples

print("dualHeatmap")
#> [1] "dualHeatmap"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)
#> Warning: executing %dopar% sequentially: no parallel backend registered

# since model training takes time and it can be reused, 
# it is advisable to save the model
# save(ds,file='...',compress='bzip2')

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
#> [1] "REACTOME-GOBP"
pairs.p <- pValuesLR(ds.LR,ds$param)
p.red.P <- reduceToPathway(pairs.p)
signatures <- getLRGeneSignatures(p.red.P,qval.thres=0.01)
scores <- scoreLRGeneSignatures(ds,signatures,rename.by.pathway=TRUE)
    
data(tme.signatures,package='BulkSignalR')
tme.scores <- scoreSignatures(ds,tme.signatures)
dualHeatmap(tme.scores,tme.scores,'example-with-TME.pdf',width=9,height=7,pointsize=4)
#> agg_png 
#>       2