Internal function sampling the (empirical) null distribution downstream the receptors

Perform receptor downstream analysis with checkReceptorSignaling based on randomized expression data and ligand-receptor pairs selected from the same randomized data.

getEmpiricalNull(
  ncounts,
  n.rand = 5,
  min.cor = -1,
  with.complex = TRUE,
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4
)

Arguments

ncounts
A matrix or table of normalized read counts.

n.rand
The number of repetitions.

min.cor
The minimum ligand-receptor correlation required.

with.complex
A logical indicating whether receptor co-complex members should be included in the target genes.

max.pw.size
Maximum pathway size to consider from the pathway reference.

min.pw.size
Minimum pathway size to consider from the pathway reference.

min.positive
Minimum number of target genes to be found in a given pathway.

Value

A list of n.rand tables such as output by checkReceptorSignaling. Each table is computed from a randomized expression matrix (randomized ncounts).

Details

A large number of correlations (ligand-receptor and receptor-downstream target genes) is reported in each table. Therefore, n.rand should be given a modest value to avoid unnecessarily long computations.

See checkReceptorSignaling for more details about the parameters.

This function is made visible to allow advanced users to develop their own statistical models or to estimate their own statistical significance.

Examples

print("getEmpiricalNull")
#> [1] "getEmpiricalNull"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds.RT.null <- getEmpiricalNull(ds$ncounts)
#> [1] "REACTOME-GOBP"
#> [1] "REACTOME-GOBP"
#> [1] "REACTOME-GOBP"
#> [1] "REACTOME-GOBP"
#> [1] "REACTOME-GOBP"