Internal function sampling of the (empirical) null distribution of ligand-receptor correlations

Perform a ligand-receptor correlation analysis based on randomized expression data.

getEmpiricalNullCorrLR(ncounts, n.rand = 5, min.cor = -1)

Arguments

ncounts
A matrix or table of normalized read counts.

n.rand
The number of repetitions.

min.cor
The minimum ligand-receptor correlation required.

Value

A list of n.rand tables such as output by getCorrelatedLR. Each table is computed from a randomized expression matrix (randomized ncounts).

Details

A large number of correlations is reported in each table. Therefore, n.rand should be given a modest value to avoid unnecessarily long computations.

See getCorrelatedLR for more details about the parameters.

This function is made visible to allow advanced users to develop their own statistical models or to estimate their own statistical significance.

Examples

print("getEmpiricalNullCorrLR")
#> [1] "getEmpiricalNullCorrLR"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds.LR.null <- getEmpiricalNullCorrLR(ds$ncounts)