Extract gene signatures of LR pair activity

Obtains gene signatures reflecting ligand-receptor as well as receptor downstream activity to score ligand-receptor pairs across samples subsequently with scoreLRGeneSignatures().

getLRGeneSignatures(pairs, pval.thres = NULL, qval.thres = NULL, signed = TRUE)

Arguments

pairs
A ligand-receptor table such as output by pValuesLR.

pval.thres
P-value threshold.

qval.thres
Q-value threshold.

signed
A logical indicating whether correlations were be considered with their signs upon P-value estimations.

Value

A table with a gene signature for each triple ligand-receptor pair. A reduction to the best pathway for each pair is automatically performed and the gene signature is comprised of the ligand, the receptor, and all the target genes with rank equal or superior to pairs$rank. In case signed==TRUE, the rank is defined for correlation absolute values.

Examples

print("getLRGeneSignatures")
#> [1] "getLRGeneSignatures"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, 
# it is advisable to save the model
# save(ds,file='...',compress='bzip2')

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
#> [1] "REACTOME-GOBP"
pairs.p <- pValuesLR(ds.LR,ds$param)
p.red.P <- reduceToPathway(pairs.p)
signatures <- getLRGeneSignatures(p.red.P,qval.thres=0.01)