Inference of ligand-receptor interactions

Computes putative LR interactions along with their statistical confidence. In this initial inference, all the relevant pathways are reported, see reduction functions to reduce this list.

# S4 method for BSRDataModel
initialInference(
  obj,
  rank.p = 0.65,
  min.cor = 0.25,
  restrict.genes = NULL,
  reference = c("REACTOME-GOBP", "REACTOME", "GOBP"),
  max.pw.size = NULL,
  min.pw.size = NULL,
  min.positive = NULL,
  restrict.pw = NULL,
  with.complex = NULL,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY",
    "ABH", "TSBH")
)

Arguments

rank.p

A number between 0 and 1 defining the rank of the last considered target genes.

min.cor

The minimum Spearman correlation required between the ligand and the receptor.

restrict.genes

A list of gene symbols that restricts ligands and receptors.

reference

Which pathway reference should be used ("REACTOME" for Reactome, "GOBP" for GO Biological Process, or "REACTOME-GOBP" for both).

max.pw.size

Maximum pathway size to consider from the pathway reference.

min.pw.size

Minimum pathway size to consider from the pathway reference.

min.positive

Minimum number of target genes to be found in a given pathway.

restrict.pw

A list of pathway IDs to restrict the application of the function.

with.complex

A logical indicating whether receptor co-complex members should be included in the target genes.

fdr.proc

The procedure for adjusting P-values according to mt.rawp2adjp.

ds

A BSRDataModel output by prepareDataset with statistical model parameters trained by its learnParameters method.

Value

A BSRInference object with initial inferences set.

Details

Perform the initial ligand-receptor inference. Initial means that reduction is applied, all the (ligand, receptor, downstream pathway) triples are reported, i.e., a given LR pair may appear multiple times with different pathways downstream the receptor. Specific reduction functions are available from the package to operate subsequent simplifications based on the BSRInference object created by the initial inference.

Parameters defining minimum/maximum pathway sizes, etc. are set to NULL by default, meaning that their values will be taken from what was set during the training of the statistical model with BSRDataModel method (learnParameters). To use different values at the time of inference sounds like a bad idea, although this could be used to explore without retraining the underlying model. Retraining of the model with adjusted parameters is advised following such an exploration.

Examples

print('prepareDataset')
#> [1] "prepareDataset"
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
#> [1] "learnParameters"
bsrdm <- learnParameters(bsrdm)
bsrdm
#> Expression values are log2-transformed: FALSE
#> Normalization method: UQ
#> Organism : hsapiens
#> Statistical model parameters:
#> $n.rand.LR
#> [1] 5
#> 
#> $n.rand.RT
#> [1] 2
#> 
#> $with.complex
#> [1] TRUE
#> 
#> $max.pw.size
#> [1] 200
#> 
#> $min.pw.size
#> [1] 5
#> 
#> $min.positive
#> [1] 4
#> 
#> $quick
#> [1] TRUE
#> 
#> $min.corr.LR
#> [1] -1
#> 
#> $LR.0
#> $LR.0$n
#> [1] 12845
#> 
#> $LR.0$model
#> $LR.0$model$mu
#> [1] 0.01456604
#> 
#> $LR.0$model$sigma
#> [1] 0.2690923
#> 
#> $LR.0$model$factor
#> [1] 0.9997934
#> 
#> $LR.0$model$start
#> [1] 8.151676e-05
#> 
#> $LR.0$model$distrib
#> [1] "censored_normal"
#> 
#> 
#> 
#> $RT.0
#> $RT.0$n
#> [1] 12845
#> 
#> $RT.0$model
#> $RT.0$model$mu
#> [1] 0.01456604
#> 
#> $RT.0$model$sigma
#> [1] 0.2690923
#> 
#> $RT.0$model$factor
#> [1] 0.9997934
#> 
#> $RT.0$model$start
#> [1] 8.151676e-05
#> 
#> $RT.0$model$distrib
#> [1] "censored_normal"
#> 
#> 
#> 
#> Expression data:
#>                SDC16     SDC15       SDC14      SDC13     SDC12     SDC11
#> TSPAN6   1313.241965 2291.4755 1685.356530 3390.62577 2629.9515 3642.8267
#> TNMD        3.452871    0.0000    2.044095   41.06738    0.0000    0.0000
#> DPM1     1322.449621 1936.3934 1613.813197 2221.91651 2505.1336 2425.3955
#> SCYL3    2087.836042  706.4846 3590.453299 1199.50980  981.9656 1176.7239
#> C1orf112  504.119177  934.6203  518.178145  983.05047  583.1335  962.7741
#> FGR       438.514626  848.1495  313.768620  219.02604  242.8098  853.9058
#>                SDC10      SDC9      SDC8     SDC7
#> TSPAN6   3060.266646  574.4433 2371.5657 2128.219
#> TNMD        9.824291    0.0000  260.2450    0.000
#> DPM1     1660.305179 1118.7139 1307.8978 3541.348
#> SCYL3    2457.055179 3574.3141 1253.1796 1268.193
#> C1orf112  455.847102  616.2210  517.8208 1159.490
#> FGR       298.658446  411.9745  547.1817 1123.256

print('perform inference')
#> [1] "perform inference"
bsrinf <- initialInference(bsrdm)
bsrinf
#> Reference database: REACTOME-GOBP
#>           L     R         pval         qval        pw.id
#> 1722    B2M  CD3D 4.192622e-35 1.892203e-31 R-HSA-202403
#> 12421 HLA-C  CD3D 7.473157e-35 1.892203e-31 R-HSA-202403
#> 12351 HLA-B  CD3D 1.149697e-34 1.940689e-31 R-HSA-202403
#> 12261 HLA-A  CD3D 1.614128e-34 2.043487e-31 R-HSA-202403
#> 1716    B2M CD247 7.333989e-33 6.189887e-30   GO:0050852
#> 2314   BTLA CD247 6.990944e-33 6.189887e-30   GO:0050852
#>                                 pw.name
#> 1722                      TCR signaling
#> 12421                     TCR signaling
#> 12351                     TCR signaling
#> 12261                     TCR signaling
#> 1716  T cell receptor signaling pathway
#> 2314  T cell receptor signaling pathway
#> Inference parameters:
#> $min.corr
#> [1] 0.25
#> 
#> $reference
#> [1] "REACTOME-GOBP"
#> 
#> $min.pw.size
#> [1] 5
#> 
#> $max.pw.size
#> [1] 200
#> 
#> $with.complex
#> [1] TRUE
#> 
#> $min.positive
#> [1] 4
#> 
#> $fdr.proc
#> [1] "BH"
#> 
#> $rank.p
#> [1] 0.65
#> 
#> $ligand.reduced
#> [1] FALSE
#> 
#> $receptor.reduced
#> [1] FALSE
#> 
#> $pathway.reduced
#> [1] FALSE
#>