Training of BulkSignalR model parameters

Unique entry point for training the parameters behind BulkSignalR statistical models.

# S4 method for BSRDataModel
learnParameters(
  obj,
  plot.folder = NULL,
  verbose = FALSE,
  n.rand.LR = 5L,
  n.rand.RT = 2L,
  with.complex = TRUE,
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  quick = TRUE,
  null.model = c("normal", "mixedNormal", "stable"),
  FUN = NULL
)

Arguments

plot.folder

A folder name for generating control plots.

verbose

A logical activating progress messages for the user.

n.rand.LR

The number of random expression matrices to use for learning the ligand-receptor correlation distribution.

n.rand.RT

The number of random expression matrices to use for learning the receptor-target genes correlation distribution.

with.complex

A logical indicating whether receptor co-complex members should be included in the target genes.

max.pw.size

Maximum pathway size to consider from the pathway reference.

min.pw.size

Minimum pathway size to consider from the pathway reference.

min.positive

Minimum number of target genes to be found in a given pathway.

quick

A logical indicating whether approximate parameters for the receptor-target correlations should be used.

null.model

The null model to use for Spearman correlation null distributions.

Value

A BSRDataModel with trained model parameters

Details

Estimates the model parameters that are stored in the slot param.

In a reference pathway, i.e., a Reactome pathway or the genes of a GOBP term, the target genes are the genes coding for proteins forming a complex with the receptor and the genes in the pathway downstream the receptor, which are given as regulated by the pathway. If with.complex is set to FALSE, then only the regulated genes are considered. Participation to a complex, being regulated, and pathway directed topologies are defined by Reactome and KEGG pathways as provided by PathwayCommons.

The min.pw.size, max.pw.size, and min.positive parameters should be identical to the values intended when searching for ligand-receptor pairs with getCorrelatedLR and checkReceptorSignaling. Although the statistical distributions are rather robust, it is not advisable to use different parameters that could introduce unanticipated biases, but for saving compute time and exploring.

The maximum pathway size is used to limit the redundancy inherent to GOBP and Reactome. The minimum pathway size is used to avoid overspecific, noninformative results.

Examples

print('prepareDataset')
#> [1] "prepareDataset"
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
#> [1] "learnParameters"
print('prepareDataset')
#> [1] "prepareDataset"
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
#> [1] "learnParameters"
bsrdm <- learnParameters(bsrdm)
bsrdm
#> Expression values are log2-transformed: FALSE
#> Normalization method: UQ
#> Organism : hsapiens
#> Statistical model parameters:
#> $n.rand.LR
#> [1] 5
#> 
#> $n.rand.RT
#> [1] 2
#> 
#> $with.complex
#> [1] TRUE
#> 
#> $max.pw.size
#> [1] 200
#> 
#> $min.pw.size
#> [1] 5
#> 
#> $min.positive
#> [1] 4
#> 
#> $quick
#> [1] TRUE
#> 
#> $min.corr.LR
#> [1] -1
#> 
#> $LR.0
#> $LR.0$n
#> [1] 12845
#> 
#> $LR.0$model
#> $LR.0$model$mu
#> [1] 0.009343633
#> 
#> $LR.0$model$sigma
#> [1] 0.2714287
#> 
#> $LR.0$model$factor
#> [1] 0.9997686
#> 
#> $LR.0$model$start
#> [1] 0.000100152
#> 
#> $LR.0$model$distrib
#> [1] "censored_normal"
#> 
#> 
#> 
#> $RT.0
#> $RT.0$n
#> [1] 12845
#> 
#> $RT.0$model
#> $RT.0$model$mu
#> [1] 0.009343633
#> 
#> $RT.0$model$sigma
#> [1] 0.2714287
#> 
#> $RT.0$model$factor
#> [1] 0.9997686
#> 
#> $RT.0$model$start
#> [1] 0.000100152
#> 
#> $RT.0$model$distrib
#> [1] "censored_normal"
#> 
#> 
#> 
#> Expression data:
#>                SDC16     SDC15       SDC14      SDC13     SDC12     SDC11
#> TSPAN6   1313.241965 2291.4755 1685.356530 3390.62577 2629.9515 3642.8267
#> TNMD        3.452871    0.0000    2.044095   41.06738    0.0000    0.0000
#> DPM1     1322.449621 1936.3934 1613.813197 2221.91651 2505.1336 2425.3955
#> SCYL3    2087.836042  706.4846 3590.453299 1199.50980  981.9656 1176.7239
#> C1orf112  504.119177  934.6203  518.178145  983.05047  583.1335  962.7741
#> FGR       438.514626  848.1495  313.768620  219.02604  242.8098  853.9058
#>                SDC10      SDC9      SDC8     SDC7
#> TSPAN6   3060.266646  574.4433 2371.5657 2128.219
#> TNMD        9.824291    0.0000  260.2450    0.000
#> DPM1     1660.305179 1118.7139 1307.8978 3541.348
#> SCYL3    2457.055179 3574.3141 1253.1796 1268.193
#> C1orf112  455.847102  616.2210  517.8208 1159.490
#> FGR       298.658446  411.9745  547.1817 1123.256