Ligand-receptor pair P-value estimation

Estimate the P-value of each ligand-receptor pair in the merged.pairs slot of the ligand-receptor analysis lr.

pValuesLR(
  lr,
  param,
  rank.p = 0.75,
  signed = TRUE,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY",
    "ABH", "TSBH"),
  use.gamma = FALSE,
  best.pw.only = FALSE
)

Arguments

lr
A ligand-receptor analysis such as output by checkReceptorSignaling.

param
A list containing the statistical model parameters.

rank.p
A number between 0 and 1 defining the rank of the last considered target genes.

signed
A logical indicating whether correlations should be considered with their signs.

fdr.proc
The procedure for adjusting P-values according to mt.rawp2adjp.

use.gamma
A logical indicating whether the Gamma distribution should be used instead of the Gaussian.

best.pw.only
A logical indicating that only the pathway with best P-value should be reported for each ligand-receptor pair.

Value

A table reporting every ligand-receptor pair with P-values and Q-values (adjusted P-values).

Examples

print('pValuesLR')
#> [1] "pValuesLR"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused,
# it is advisable to save the model
# save(ds,file='...',compress='bzip2')

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
#> [1] "REACTOME-GOBP"
pairs.p <- pValuesLR(ds.LR,ds$param)