Prepare a BulkSignalR data set

Take a matrix or data frame containing read counts along with sample types and combine them into a list for subsequent use with BulkSignalR. Expression proteomics data could be used as well.

prepareDataset(
  counts,
  types,
  normalize = TRUE,
  symbol.col = NULL,
  min.count = 10,
  prop = 0.1,
  method = c("UQ", "TC"),
  log.transformed = FALSE
)

Arguments

counts
A table or matrix of read counts.

types
A vector of characters defining sample types.

normalize
A logical indicating whether counts should be normalized according to method or it was normalized beforehand.

symbol.col
The index of the column containing the gene symbols in case those are not the rownames of counts already.

min.count
The minimum read count of a gene to be considered expressed in a sample.

prop
The minimum proportion of samples where a must be expressed to keep that gene.

method
The normalization method ('UQ' for upper quartile or 'TC' for total count).

log.transformed
A logical indicating whether expression data were already log-transformed, e.g., some microarray data.

Value

A list containing the read counts and the sample types for further use in BulkSignalR.

Details

The counts matrix or table should be provided with expression levels of protein coding genes in each samples (column) and rownames(counts) set to the gene symbols. For commodity, it is also possible to provide the expression matrix or table counts with the gene symbols stored in one of the columns. This column must be specified with symbol.col. In such a case, prepareDataset will extract this column and use it to set the row names. Because row names must be unique, prepareDataset will eliminate rows with duplicated gene symbols by keeping the rows with maximum average expression. Gene symbol duplication typically occurs in protein coding genes after genome alignment due to errors in genome feature annotation files (GTF/GFF), where a handful of deprecated gene annotations may remain along with the current ones, or some genes are not given their fully specific symbols. If your read count extraction pipeline does not take care of this phenomenon, the maximum mean expression selection strategy implemented here should solve this difficulty for the sake of inferring ligand-receptor interactions.

If normalize is TRUE then normalization is performed according to method. If those two simple methods are not satisfying, then it is possible to provide a pre-normalized matrix, e.g., using edgeR TMM algorithm.

In case proteomic data are provided, min.count must be understood as its equivalent with respect to those data (spectral counts, iBAQ, etc.).

Examples

print('prepareDataset')
#> [1] "prepareDataset"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)