Aggregate ligands and receptors at the pathway level

Simplifies a ligand-receptor table to focus on the pathways.

reduceToPathway(pairs)

Arguments

pairs
A ligand-receptor table such as output by pValuesLR.

Value

A table with the same structure as pairs but reduced to only report one row per pathway.

Details

All the ligands and all the receptors forming pairs related to a certain pathway, are combined in two semi-colon-separated lists surrounded by curly braces. The information of which ligand interacted with which receptor is lost. For a given pathway, the reported P-values and target genes are those of the best ligand-receptor pair that was in this pathway.

There is no pathway selection, i.e. potentially redundant pathways with different P-values are all kept in the output. reduceToBestPathway() can be used to select the one with best P-value.

Examples

print('reduceToPathway')
#> [1] "reduceToPathway"
data(sdc,package='BulkSignalR')
sample.types <- rep('tumor',ncol(sdc))
sample.types[grep('^N',names(sdc),perl=TRUE)] <- 'normal'
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
# save(ds,file='...',compress='bzip2')

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
#> [1] "REACTOME-GOBP"
pairs.p <- pValuesLR(ds.LR,ds$param)
pairs.pw <- reduceToPathway(pairs.p)
pairs.pw.red <- reduceToBestPathway(pairs.pw)