Score ligand-receptor gene signatures

Compute ligand-receptor gene signature scores over a BSRDataModel.

# S4 method for BSRDataModel
scoreLRGeneSignatures(
  obj,
  sig,
  LR.weight = 0.5,
  robust = FALSE,
  name.by.pathway = FALSE,
  abs.z.score = FALSE
)

Arguments

sig

A BSRSignature object.

LR.weight

A number between 0 and 1 defining the relative weight of the ligand and the receptor in the signature.

robust

A logical indicating that z-scores should be computed with median and MAD instead of mean and standard deviation.

name.by.pathway

A logical indicating whether row names of the resulting score matrix should be complemented with (illustrative) pathway names.

abs.z.score

A logical to use absolute z-scores (useful if the activity of a paythway is reported by a mixture of up- and down-genes whose z-score averages might hide actual activity).

Value

A matrix containing the scores of each ligand-receptor gene signature in each sample.

Examples

if(FALSE){}