Heatmap function for LR scores and additional data

Generate a stack of heatmaps. The top heatmap represents ligand-receptor gene signature scores, while the bottom heatmap represents the second, user-chosen data set to feature along with gene signatures.

Usage

dualHeatmap(
  mat.c,
  mat.e,
  width,
  height,
  file.name = NULL,
  dend.row = NULL,
  dend.spl = NULL,
  dend.e = NULL,
  cols = NULL,
  cols.e = NULL,
  pointsize = 4,
  cut.p = 0.01,
  vert.p = 0.9,
  row.names.1 = TRUE,
  row.names.2 = TRUE
)

Arguments

mat.c

A matrix with the signature scores such as output by scoreLRGeneSignatures.

mat.e

A second matrix to feature below mat.c.

width

PDF width.

height

PDF height.

file.name

A PDF file name.

dend.row

A precomputed row dendrogram for mat.c.

dend.spl

A precompute sample (column) dendrogram.

dend.e

A precomputed row dendrogram for mat.e.

cols

A vector of colors to use for the heatmap of mat.c.

cols.e

A vector of colors to use for the heatmap of mat.e.

pointsize

PDF pointsize.

cut.p

Proportion of top and bottom values for thresholding.

vert.p

Proportion of total height dedicated to the top heatmap.

row.names.1

A logical to turn on/off the display of row (gene) names in the top heatmap.

row.names.2

A logical to turn on/off the display of row (gene) names in the bottom heatmap.

Value

A heatmap. Since heatmap plotting tend to be slow on the screen, it is advisable to provide a PDF file name and plot in a file (much faster).

Extreme values (top and bottom) can be replaced by global quantiles at cut.p and 1-cut.p

to avoid color scales shrunk by a few outliers.

This is a convenience function that relies on the ComplexHeatmap

package to propose a simple way of representing signature scores. If more advanced features are needed or more graphic parameters should be controlled, users should implement their own function.

Examples

print('simpleHeatmap')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redP <- reduceToPathway(bsrinf) 
bsrinf.redPBP <- reduceToBestPathway(bsrinf.redP) 
bsrsig.redPBP <- getLRGeneSignatures(bsrinf.redPBP,
qval.thres=0.001)

scoresPathway <- scoreLRGeneSignatures(bsrdm,
               bsrsig.redPBP,name.by.pathway=TRUE)

data(immune.signatures, package="BulkSignalR")
imm.scores <- scoreSignatures(bsrdm, immune.signatures)
dualHeatmap(scoresPathway, imm.scores, width=6, height=9,
           file="SDC-LR-dualheatmap.pdf",
pointsize=4, vert.p=0.85)