Inference of ligand-receptor interactions

Computes putative LR interactions along with their statistical confidence. In this initial inference, all the relevant pathways are reported, see reduction functions to reduce this list.

Usage

# S4 method for BSRDataModel
initialInference(
  obj,
  rank.p = 0.55,
  min.cor = 0.25,
  restrict.genes = NULL,
  reference = c("REACTOME-GOBP", "REACTOME", "GOBP"),
  max.pw.size = NULL,
  min.pw.size = NULL,
  min.positive = NULL,
  restrict.pw = NULL,
  with.complex = NULL,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY",
    "ABH", "TSBH")
)

Arguments

rank.p

A number between 0 and 1 defining the rank of the last considered target genes.

min.cor

The minimum Spearman correlation required between the ligand and the receptor.

restrict.genes

A list of gene symbols that restricts ligands and receptors.

reference

Which pathway reference should be used ("REACTOME" for Reactome, "GOBP" for GO Biological Process, or "REACTOME-GOBP" for both).

max.pw.size

Maximum pathway size to consider from the pathway reference.

min.pw.size

Minimum pathway size to consider from the pathway reference.

min.positive

Minimum number of target genes to be found in a given pathway.

restrict.pw

A list of pathway IDs to restrict the application of the function.

with.complex

A logical indicating whether receptor co-complex members should be included in the target genes.

fdr.proc

The procedure for adjusting P-values according to mt.rawp2adjp.

ds

A BSRDataModel output by prepareDataset with statistical model parameters trained by its learnParameters method.

Value

A BSRInference object with initial inferences set.

Details

Perform the initial ligand-receptor inference. Initial means that reduction is applied, all the (ligand, receptor, downstream pathway) triples are reported, i.e., a given LR pair may appear multiple times with different pathways downstream the receptor. Specific reduction functions are available from the package to operate subsequent simplifications based on the BSRInference object created by the initial inference.

Parameters defining minimum/maximum pathway sizes, etc. are set to NULL by default, meaning that their values will be taken from what was set during the training of the statistical model with BSRDataModel method (learnParameters). To use different values at the time of inference sounds like a bad idea, although this could be used to explore without retraining the underlying model. Retraining of the model with adjusted parameters is advised following such an exploration.

Examples

print('prepareDataset')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
bsrdm <- learnParameters(bsrdm)
bsrdm

print('perform inference')
bsrinf <- initialInference(bsrdm)
bsrinf