Prepare a BSRDataModel object from expression data

Take a matrix or data frame containing RNA sequencing, microarray, or expression proteomics data and returns a BSRDataModel object ready for subsequent training. Normally, BSRDataModel objects are not instantiated directly, but through this function.

prepareDataset(
  counts,
  normalize = TRUE,
  symbol.col = NULL,
  min.count = 10,
  prop = 0.1,
  method = c("UQ", "TC"),
  log.transformed = FALSE,
  min.LR.found = 80,
  species = "hsapiens",
  conversion.dict = data.frame(Gene.name = "A", row.names = "B"),
  UQ.pc = 0.75
)

Arguments

counts

A table or matrix of read counts.

normalize

A logical indicating whether counts should be normalized according to method or if it was normalized beforehand.

symbol.col

The index of the column containing the gene symbols in case those are not the row names of counts already.

min.count

The minimum read count of a gene to be considered expressed in a sample.

prop

The minimum proportion of samples where a gene must be expressed higher than min.count to keep that gene.

method

The normalization method ('UQ' for upper quartile or 'TC' for total count).

log.transformed

A logical indicating whether expression data were already log2-transformed, e.g., some microarray data.

min.LR.found

The minimum number of ligands or receptors found in count row names after eliminating the rows containing too many zeros according to min.count and prop.

UQ.pc

Percentile for upper-quartile normalization, number between 0 and 1 (in case the default 0.75 - hence the name - is not appropriate).

organism

Organism

Value

A BSRModelData object with empty model parameters.

Details

The counts matrix or table should be provided with expression levels of protein coding genes in each samples (column) and rownames(counts) set to HUGO offcial gene symbols. For commodity, it is also possible to provide counts with the gene symbols stored in one of its columns. This column must be specified with symbol.col. In such a case, prepareDataset will extract this column and use it to set the row names. Because row names must be unique, prepareDataset will eliminate rows with duplicated gene symbols by keeping the rows with maximum average expression. Gene symbol duplication may occur in protein coding genes after genome alignment due to errors in genome feature annotation files (GTF/GFF), where a handful of deprecated gene annotations might remain, or some genes are not given their fully specific symbols. If your read count extraction pipeline does not take care of this phenomenon, the maximum mean expression selection strategy implemented here should solve this difficulty for the sake of inferring ligand-receptor interactions.

If normalize is TRUE then normalization is performed according to method. If those two simple methods are not satisfying, then it is possible to provide a pre-normalized matrix setting normalize to FALSE.

In case proteomic or microarray data are provided, min.count must be understood as its equivalent with respect to those data.

Examples

print('prepareDataset')
#> [1] "prepareDataset"
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])